requirements.txt
asgiref==3.5.2
Django==4.1.3
django-restframework==0.0.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
